prepare for deploy

+++ start_env.sh
@@ -82,7 +82,7 @@ fi
 
 if [ $1 = "pupp" ];
 then
-  DOCKERFILE="Dockerfile_puppeteer"
+  # DOCKERFILE="Dockerfile_puppeteer"
   CONTAINER_PORT=5000
 fi
 
@@ -102,14 +102,17 @@ then
   DOCKERFILE="Dockerfile_dev"
 fi
 
-echo
-echo "${bold}${cyan}================= Building container ===================${reset}"
-cp containers/$DOCKERFILE Dockerfile
+if [ $1 != "pupp" ];
+then
+  echo
+  echo "${bold}${cyan}================= Building container ===================${reset}"
+  cp containers/$DOCKERFILE Dockerfile
 
-GUNICORN_PORT=4000
-docker build -t devenv-$1 .
+  GUNICORN_PORT=4000
+  docker build -t devenv-$1 .
 
-rm Dockerfile
+  rm Dockerfile
+fi
 
 # --env-file=$PROJECT_PATH/containers/env.txt \
 echo
@@ -132,7 +135,7 @@ then
     -v $PROJECT_PATH/containers/pupp/jest.config.js:/home/pptruser/jest.config.js \
     -v $PROJECT_PATH/containers/pupp/jest-puppeteer.config.js:/home/pptruser/jest-puppeteer.config.js \
     -v $PROJECT_PATH/.babelrc:/home/pptruser/.babelrc \
-    devenv-$1 \
+    airladon/pynode:python3.7.2-node10.15.0-npm6.6.0-puppeteer \
     bash
 else
   # docker volume create browser-tests
